Replace the local shouldEdit flag with a context-level array of flags
To avoid having the old CatCard component still displayed in the HotspotDetailScreen after an update, the local state "shouldEdit" is now deleted and a context-level array is now initialized for each type of cat (sterilized/unsterilized).
This commit should be squashed together with the previous ones.
TODO: When exiting the HotspotFormScreen after modifying cat details without saving the hotspot, both the old and the new CatCard will be displayed in the HotspotForm when accessing it once more.