Fix null email and wrongly formatted check-in/out date issue
The objToFormData function used to fail when updating a particular cat card due to the e-mail field containing a null value. Also, the dates selected in the calendar were not properly formatted as a YYYY-MM-DD date string (which seemed to be why the date was not updated/refreshed).

The objToFormData should obviously be refactored, but the nested ifs are probably created to handle conditions specifically for the media updates and they could be removed or somehow refactored if any types were not used.