Replace forward/reverse geocoding service with geocode.maps.co
geocode.maps.co is a free geocoding service that offers 1 million free requests per month, far more than MapBox's free-tier account.
TODO: Try to fix the Marker-related error displayed in the browser's console for the web platform. It seems that this issue slows down the selection of new locations in the ChooseLocationScreen.