Fix the cat paw markers on web
As expected, the issue was the relative path used to refer to the images supposed to be displayed as marker pins. The web platform creates yet another "assets/" directory where the project's "assets" directory is copied, hence why the app did not display the paw markers.

Additionally, we have a separate Marker component which is a bit more convenient.